+++ app/serializers/auth.go
@@ -4,3 +4,7 @@ type LoginReq struct {
 	Email    string `json:"email"`
 	Password string `json:"password"`
 }
+
+type RefreshTokenReq struct {
+	RefreshToken string `json:"refresh_token"`
+}