+++ app/svc/auth.go
@@ -7,4 +7,5 @@ import (
 
 type IAuth interface {
 	Login(req *serializers.LoginReq) (interface{}, *errors.RestErr)
+	RefreshToken(req *serializers.RefreshTokenReq) (interface{}, *errors.RestErr)
 }