+++ infra/config/config.go
@@ -21,6 +21,7 @@ type FireBaseConfig struct {
 	ApiKey                        string
 	SignUpWithEmailAndPasswordUrl string
 	SignInWithEmailAndPasswordUrl string
+	RefreshTokenUrl               string
 	Timeout                       time.Duration
 }
 
@@ -80,9 +81,10 @@ func setDefaultConfig() {
 
 	config.FireBase = &FireBaseConfig{
 		CredentialFilePath:            "fb-svc-key.json",
-		ApiKey:                        "AIzaSyBWzra2RfPaym9Jvm69CtjRl5y-5LjsI90",
+		ApiKey:                        "fb-web-api-key",
 		SignUpWithEmailAndPasswordUrl: "https://identitytoolkit.googleapis.com/v1/accounts:signUp",
 		SignInWithEmailAndPasswordUrl: "https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword",
+		RefreshTokenUrl:               "https://securetoken.googleapis.com/v1/token",
 		Timeout:                       10,
 	}
 }